+++ src/components/comment-form.tsx
@@ -1,67 +1,37 @@
 import { useState } from 'react';
-import React from 'react';
+import Rating from './rating';
 
-const STARS_NUMBER = 5;
-const STAR_TITLES = ['terribly', 'badly', 'not bad', 'good', 'perfect'] as const;
 const TEXT_MIN_LENGTH = 50;
 const TEXT_MAX_LENGTH = 300;
 
+type Form = HTMLFormElement & {
+  rating: RadioNodeList;
+  review: HTMLTextAreaElement;
+}
+
 function CommentForm(): JSX.Element {
-  const [formData, setFormData] = useState({ rating: 0, review: '' });
+  const [isSubmitDisabled, setSubmitDisabled] = useState(true);
 
-  const handleFieldChange = (evt: React.ChangeEvent<HTMLInputElement> | React.ChangeEvent<HTMLTextAreaElement>) => {
-    const { name, value } = evt.target;
-    setFormData({ ...formData, [name]: value });
-  };
-
-  function disableSubmit() {
-    return formData.review.length <= TEXT_MIN_LENGTH || formData.rating === 0;
-  }
-
-  function renderStars() {
-    const stars: JSX.Element[] = [];
-    for (let i = STARS_NUMBER; i > 0; i--) {
-      stars.push(
-        <React.Fragment key={`star-${i}`}>
-          <input
-            className="form__rating-input visually-hidden"
-            name="rating"
-            defaultValue={i}
-            id={`${i}-stars`}
-            type="radio"
-            onChange={handleFieldChange}
-            defaultChecked={formData.rating === i}
-          />
-          <label
-            htmlFor={`${i}-stars`}
-            className="reviews__rating-label form__rating-label"
-            title={STAR_TITLES[i - 1]}
-          >
-            <svg className="form__star-image" width={37} height={33}>
-              <use xlinkHref="#icon-star" />
-            </svg>
-          </label>
-        </React.Fragment>
-      );
-    }
-    return stars;
+  function handleFormChange(evt: React.FormEvent<HTMLFormElement>) {
+    const form = evt.currentTarget as Form;
+    const rating = form.rating.value;
+    const review = form.review.value;
+    setSubmitDisabled(review.length <= TEXT_MIN_LENGTH || !rating);
   }
 
   return (
-    <form className="reviews__form form" action="#" method="post">
+    <form className="reviews__form form" action="#" method="post" onChange={handleFormChange}>
       <label className="reviews__label form__label" htmlFor="review">
         Your review
       </label>
       <div className="reviews__rating-form form__rating">
-        {renderStars()}
+        <Rating />
       </div>
       <textarea
         className="reviews__textarea form__textarea"
         id="review"
         name="review"
         placeholder="Tell how was your stay, what you like and what can be improved"
-        value={formData.review}
-        onChange={handleFieldChange}
         maxLength={TEXT_MAX_LENGTH}
       />
       <div className="reviews__button-wrapper">
@@ -73,7 +43,7 @@ function CommentForm(): JSX.Element {
         <button
           className="reviews__submit form__submit button"
           type="submit"
-          disabled={disableSubmit()}
+          disabled={isSubmitDisabled}
         >
           Submit
         </button>
